feat: changed configVersion to be a double value instead of an int value

+++ src/main/java/com/dragonminez/common/config/CombatConfig.java
@@ -15,10 +15,10 @@ import java.util.Map;
 @Getter
 @NoArgsConstructor
 public class CombatConfig {
-	public static final int CURRENT_VERSION = ConfigManager.CONFIG_VERSION;
+	public static final double CURRENT_VERSION = ConfigManager.CONFIG_VERSION;
 
 	@Setter
-	private int configVersion;
+	private double configVersion;
 
 	private Double staminaConsumptionRatio = 0.083;
 	private Double blockStaminaCost = 0.25;
